steering_vector_factory/requirements.txt
# Steering Vector Factory Dependencies
# Separate from SAM's main requirements to avoid conflicts

# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Data processing
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0

# API access for teacher model
openai>=1.3.0
anthropic>=0.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
jsonlines>=3.1.0

# Optional: For model optimization
bitsandbytes>=0.41.0
